It works now! Cleaned up and added reasonable values.

+++ src/character.h
@@ -3000,24 +3000,24 @@ class Character : public Creature, public visitable
         void mod_rad( int mod );
 
         float get_heartrate_index() const;
-        void set_heartrate_effect_mod( int  mod);
-        void modify_heartrate_effect_mod( int mod);
-        float get_heartrate_effect_mod() const;
+        void set_heartrate_effect_mod( int  mod );
+        void modify_heartrate_effect_mod( int mod );
+        int get_heartrate_effect_mod() const;
         void update_heartrate_index();
 
         float get_bloodvol_index() const;
         void update_bloodvol_index();
 
         float get_circulation_resistance() const;
-        void set_bp_effect_mod(int mod);
-        void modify_bp_effect_mod(int mod);
-        float get_bp_effect_mod() const;
+        void set_bp_effect_mod( int mod );
+        void modify_bp_effect_mod( int mod );
+        int get_bp_effect_mod() const;
         void update_circulation_resistance();
 
         float get_perspiration_rate() const;
-        void set_perspiration_effect_mod( int mod);
-        float get_perspiration_effect_mod() const;
-        void modify_perspiration_effect_mod( int mod);
+        void set_perspiration_effect_mod( int mod );
+        int get_perspiration_effect_mod() const;
+        void modify_perspiration_effect_mod( int mod );
         void update_perspiration_rate();
 
         void update_circulation();
